:package: refactor: home 패키지 구조 변경 및 일부 포맷팅
Issue #9
